Dashboard Design Overhaul Performance Metrics component is moved into the Components folder
